Text encryption is implemented

+++ src/text_rw.h
@@ -5,9 +5,9 @@
 #include <Windows.h>
 #include "text_data.h"
 
-void write_data_for_text_encryption_to_file(unsigned long textLength, unsigned int lowerBitsNum, char *decryptionDataFilename);
+void write_data_for_text_encryption_to_file(DWORD textLength, UINT lowerBitsNum, char *decryptionDataFilename);
 void read_data_to_encrypt_text_from_file(char *decryptionDataFilename, TEXTDATA *decryptionData);
 void read_text_for_encryption_from_file(char *pathToFile, TEXTDATA *encryptedText);
-// void write_text(char *pathToFile, unsigned char *textToWrite, unsigned long *charCount);
+// void write_decrypted_text_to_file(char *pathToFile, unsigned char *textToWrite, DWORD *charCount);
 
 #endif
